test(checkbox): migrate legacy basic
