+++ core/src/components/checkbox/test/legacy/basic/checkbox.e2e.ts-snapshots/checkbox-basic-md-rtl-Mobile-Safari-linux.png
