+++ core/src/components/checkbox/test/legacy/basic/checkbox.e2e.ts-snapshots/checkbox-legacy-basic-ios-ltr-Mobile-Safari-linux.png
