+++ core/src/components/checkbox/test/legacy/basic/checkbox.e2e.ts-snapshots/checkbox-legacy-basic-ios-rtl-Mobile-Chrome-linux.png
