+++ core/src/components/checkbox/test/legacy/basic/checkbox.e2e.ts-snapshots/checkbox-legacy-basic-md-ltr-Mobile-Firefox-linux.png
